requirements.txt
huggingface_hub>=0.21.1
fastapi
uvicorn
python-multipart
sqlalchemy
langchain
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
huggingface_hub>=0.21.1
streamlit
pytest
pypdf
faiss-cpu
python-dotenv

pytest
pytest-asyncio
httpx
